[meta] Add and use TOPDIR and BUILD_DIR arguments
Buildroot sets up its core variables differently depending
on whether you're putting output in a custom location; e.g.
when you run

    $ make O=<other location>

This caused our paths to break, as it was developed always
using a custom output directory.

This change fixes up our usage by being very explicit about
what we need and what we expect.

- For the command line, these are now the options:

  -B IDIR, --base IDIR  Buildroot Source Directory
  -o ODIR, --output ODIR
                        Buildroot Output Directory
  -b BDIR, --build BDIR
                        Buildroot Build Directory

.. which make a distinction between the Output and Build directories,
and adds the Source directory.

* If Source is unset, it defaults to the current directory.

* If Output is unset, it defaults to <Source>/output/

* If Build it unset, it defaults to <Output>/build

- external.mk sets all three:
  * Source == $(CANONICAL_CURDIR)
  * Output == $(CANONICAL_O)
  * Build == $(BUILD_DIR)

- In addition, this makes us smarter about the top-level .config

If you don't use an external output directory, the Buildroot .config
is in the top-level directory. We could add another option to
capture $(CONFIG_DIR), but instead we check for it in a few
different locations -- <Output>, <Source> and the Current Directory.
The first one found wins.