[kernel/u-boot] Add option to omit EXTRAVERSION; have U-Boot do that.
This adds a flag to _get_version_from_makefile() to control whether
to include it in the version string returned (set to 'True') by default.

EXTRAVERSION is the U-Boot Makefile is extraneous and only causes a
fixup on the server side for CVE_VERSION. So, it sets the flag to 'False'
when calling.