[packages] Fixup package names to use the 'rawname' if different
Nearly all packages delimit their names with '-'. But some use '_',
which wasn't compensated for.

This commit rectifies that by first grabbing the 'RAWNAME' from the
make variables of the package and comparing it against the name we're
using. If it differs, then we drop the current sub-dictionary and add
a new sub-dictionary with the fixed up name (and fields thereof).