[internal] Improve messaging helpers and fixup users.
This aligns the dbg()/info()/warn()/err() helpers so that they
can be called by any routine and do not need the master 'vgls'
dictionary; and so that they are formatted nicely with a leading
tag of the type of message and our name (e.g. 'INFO Vigiles:').

This commit also fixes up the current users of the helpers if
they needed it.