[features] Add support for amending CVE manifest + report
This adds support for specifying and parsing user-supplied CSV files
that modify the Vigiles manifest that is submitted for CVE checking
and hence the report that is generated for the following items:

* Additional Packages to check against

* Packages to exclude from the check

* Whitelisted CVEs to omit from the report

The changes are represented in the following:

* Kconfig options to specify the files

* Makefile support for checking and passing those to vigiles-buildroot

* Command-line options for each

* Internal helpers to open, read + process each

The internal changes are in a new file -- amendments.py and called
by a new function 'amend_manifest()' by 'write_manifest()'.

amendments.py also now houses the initial whitelist construction,
which is built from each package's 'ignore-cves' field.