[parser] Split package makefile variable parsing into 2 steps.
The variable list we were using (br2_pkg_var_list) is now split into 2 --
'br2_internal_pkg_vars' and 'br2_user_pkg_vars', where the first contains the
automatically set variables by the build system and the other the variables
that can be set/over-ridden by the package makefiles (and they've both been
alphabetized for readability and maintainability).

They are now gathered in 2 loops so that the first primes the list of known
packages, and the 2nd checks if the package is already known before setting.

This is done in case there are conflicts where the name of one variable is
a subset of another and separated with the only delimeter we can use ('_'
in makefiles, transformed to '-' in the code), and because we don't want to
rely on the order of the list of variables above for priority or the order of
the incoming strings to parse.

Think 'foo-version' vs. 'foo-cve-version' -- if 'version' is in the list
first, but 'foo-cve-version' is parsed first, then a bogus 'foo-cve' package
would be added.

Sure, if everything is guaranteed to be in alphabetical order, that particular
example is moot, but will that always be the case for what we need to parse
(i.e. it's 'future-proof'), and for how lists are internally stored?