[credentials] Add ability to override via environment.
This feature updates the semantics in handling the config variables
BR2_EXTERNAL_VIGILES_KEY_FILE and BR2_EXTERNAL_VIGILES_DASHBOARD_CONFIG.

Their values are now considered to be the default locations of those
files which can be over-ridden on the command line by setting the shell
environment variables VIGILES_KEY_FILES and VIGILES_DASHBOARD_CONFIG
respectively.

This is intended to allow developers to share a common defconfig for
a board, with default credentials for e.g. an official build server,
but use their own private/local credentials without having to manually
modify the board config for each build.