[llapi] Add verbose and non-fatal error reporting.
This improves the messaging and handling of communication errors
between the Vigiles Desktop client (this) and the LinuxLink servers.

The HTTP and URL exceptions are (lightly) parsed to capture the
error code returned so a more descriptive message can be displayed
to the user, along with some diagnostic information and instructions
should the condition be escalated into a support ticket.