[reports] Add support to upload Manifest without waiting for CVE Report
This adds a new Kconfig option (BR2_EXTERNAL_VIGILES_UPLOAD_ONLY) that can be
set to enable this feature.

This can speed up build times and ease reporting of automated bulk builds --
Instead of a text report and a link to the online report, a link to the
Vigiles Dashboard Product Workspace (as specified with
VIGILES_DASHBOARD_CONFIG) will be displayed, from where a CVE report can
be generated.