[make] Add workaround for Make 4.3 bug
This is an automatic fixup to work around the Make bug and the Buildroot
change to disable the 'printvars' target (which we fundamentally rely on).
It ain't great, but it ought to work until the upstream bug is fixed.

It checks the Make version and branches if the buggy version is found; then..

- Backs up the current Makefile
- Opens the Makefile and removes the dependencies from the 'printvars' target
  (then writes)
- Calls 'make printvars' as usual, but this time as the COMMAND argument
  to 'prlimit' after increasing the RLIMIT stack size for the process
- Moves the backup Makefile back in place.

To support this, the actual make execution was moved to a short encapsulating
function that accepts lists of arguments for subprocess.check_output():

- mk_opts (output directory, etc)
- mk_args (e.g. target)
- mk_context (preamble for e.g. wrapping command or environment changes)